Moment fuel calculation bug fix #1

+++ app/src/main/java/com/demo/carspends/presentation/fragments/noteFilling/NoteFillingAddOrEditViewModel.kt
@@ -1,6 +1,7 @@
 package com.demo.carspends.presentation.fragments.noteFilling
 
 import android.app.Application
+import android.util.Log
 import androidx.lifecycle.viewModelScope
 import com.demo.carspends.R
 import com.demo.carspends.domain.car.usecases.EditCarItemUseCase
@@ -188,7 +189,7 @@ class NoteFillingAddOrEditViewModel @Inject constructor(
                 val mileage = abs(sorted.first().mileage - sorted[1].mileage)
                 val fuel = sorted.first().liters
                 if (fuel <= 0 || mileage <= 0) 0.0
-                else fuel / (mileage / 100)
+                else fuel / (mileage.toDouble() / 100)
             } else 0.0
 
             carItem = itCar.copy(
